LayoutTests/ietestcenter/Javascript/15.3.4.5-0-2-expected.txt

15.3.4.5-0-2 (TestCases/chapter15/15.3/15.3.4/15.3.4.5/15.3.4.5-0-2.js) Function.prototype.bind must exist as a function taking 1 parameter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
